+++ Mods/QudJP/Assemblies/QudJP.Tests/DummyTargets/DummyOtherUiSharedHelpers.cs
@@ -1,0 +1,13 @@
+namespace QudJP.Tests.DummyTargets;
+
+internal sealed class DummyUiIconWithGameObject
+{
+    public DummyActiveObject gameObject = new DummyActiveObject();
+
+    public object? LastRenderable { get; private set; }
+
+    public void FromRenderable(object? renderable)
+    {
+        LastRenderable = renderable;
+    }
+}
